apps/mcp-server/pyproject.toml
[project]
name = "envel_mcp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["fastmcp", "httpx", "python-dotenv", "sqlcipher3-wheels"]

[project.scripts]
envel = "envel_mcp.server:main"

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
